requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
tinydb>=4.8.0
pyyaml>=6.0.1
wakeonlan>=3.1.0
apscheduler>=3.10.4
psutil>=5.9.6
aiofiles>=23.2.0
pydantic>=2.5.0
jinja2>=3.1.2
python-multipart>=0.0.6